chore: make types initialize attrs type same to RPC
Add more comment to types initialize to make it more clear
