+++ lib/ckb/types/types.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative "seal"
 require_relative "block_header"
 require_relative "block"
 require_relative "uncle"
